backend/requirements.txt
django
channels
djangorestframework
djangorestframework-simplejwt
django-cors-headers
daphne
Pillow
